#14032: move new tests with typecast test
